Adapt VM exec_push handler to i16 immediate
PUSH now carries an i16 field (not i64) following the bytecode refactor.
Update exec_push to accept i16 and widen to i64 before pushing onto the stack.

+++ crates/vm/src/vm.rs
@@ -480,8 +480,8 @@ impl Vm {
 
     // -- Stack & register I/O --
 
-    fn exec_push(&mut self, _pos: usize, imm: i64) -> Result<StepResult, Error> {
-        self.push_val(imm);
+    fn exec_push(&mut self, _pos: usize, imm: i16) -> Result<StepResult, Error> {
+        self.push_val(i64::from(imm));
         Ok(StepResult::Continue)
     }
